Move service/common/api/service -> service/common/httperror
