+++ internal/pkg/service/common/httperror/badrequest.go
@@ -1,22 +1,9 @@
-package service
+// Package apierror provides basic errors for all APIs.
+package httperror
 
 import (
 	"net/http"
 )
-
-type NotImplementedError struct{}
-
-func (NotImplementedError) ErrorName() string {
-	return "notImplemented"
-}
-
-func (NotImplementedError) Error() string {
-	return "operation not implemented"
-}
-
-func (NotImplementedError) ErrorUserMessage() string {
-	return "Operation not implemented."
-}
 
 type BadRequestError struct {
 	Message string
